// WebClient/src/api/GeneratedApiClient.ts
export enum AvatarGenderTypeEnum {
    Male = "Male",
    Female = "Female",
}

export interface ChatRecord {
    message?: string | undefined;
    promptType?: PromptTypeEnum;
    date?: Date;
}

export interface ChatSession {
    records?: ChatRecord[] | undefined;
}
// ...
export enum PromptTypeEnum {
    User = "User",
    Agent = "Agent",
}
// ...
export interface User {
    id?: number;
    name?: string | undefined;
    avatarGenderType?: AvatarGenderTypeEnum;
}